+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `--include-robots` flag - Extract URLs from robots.txt files by [@Adesoji1](https://github.com/Adesoji1)
 - Added `--include-sitemap` flag - Extract URLs from sitemap
+- Added `--strict` flag - Enforce exact host validation (default is true)
 
 ## 0.4.0